backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
sqlparse
tzdata
PyJWT
requests